+++ crates/ruma-common/src/api/metadata.rs
@@ -1,6 +1,6 @@
 use std::{
     cmp::Ordering,
-    fmt::{self, Display, Write},
+    fmt::{Display, Write},
     str::FromStr,
 };
 
@@ -624,6 +624,33 @@ impl MatrixVersion {
         self >= other
     }
 
+    /// Get a string representation of this Matrix version.
+    ///
+    /// This is the string that can be found in the response to one of the `GET /versions`
+    /// endpoints. Parsing this string will give the same variant.
+    ///
+    /// Returns `None` for [`MatrixVersion::V1_0`] because it can match several per-API versions.
+    pub const fn as_str(self) -> Option<&'static str> {
+        let string = match self {
+            MatrixVersion::V1_0 => return None,
+            MatrixVersion::V1_1 => "v1.1",
+            MatrixVersion::V1_2 => "v1.2",
+            MatrixVersion::V1_3 => "v1.3",
+            MatrixVersion::V1_4 => "v1.4",
+            MatrixVersion::V1_5 => "v1.5",
+            MatrixVersion::V1_6 => "v1.6",
+            MatrixVersion::V1_7 => "v1.7",
+            MatrixVersion::V1_8 => "v1.8",
+            MatrixVersion::V1_9 => "v1.9",
+            MatrixVersion::V1_10 => "v1.10",
+            MatrixVersion::V1_11 => "v1.11",
+            MatrixVersion::V1_12 => "v1.12",
+            MatrixVersion::V1_13 => "v1.13",
+        };
+
+        Some(string)
+    }
+
     /// Decompose the Matrix version into its major and minor number.
     pub const fn into_parts(self) -> (u8, u8) {
         match self {
@@ -766,13 +793,6 @@ impl MatrixVersion {
     }
 }
 
-impl Display for MatrixVersion {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let (major, minor) = self.into_parts();
-        f.write_str(&format!("v{major}.{minor}"))
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use assert_matches2::assert_matches;
@@ -898,5 +918,16 @@ mod tests {
         const LIT: MatrixVersion = MatrixVersion::from_lit("1.0");
 
         assert_eq!(LIT, V1_0);
+    }
+
+    #[test]
+    fn parse_as_str_sanity() {
+        let version = MatrixVersion::try_from("r0.5.0").unwrap();
+        assert_eq!(version, V1_0);
+        assert_eq!(version.as_str(), None);
+
+        let version = MatrixVersion::try_from("v1.1").unwrap();
+        assert_eq!(version, V1_1);
+        assert_eq!(version.as_str(), Some("v1.1"));
     }
 }